backend/requirements/development.txt
-r base.txt
pytest-django
model-bakery
black
isort
flake8
django-debug-toolbar
